restructuring project arquitecture
