+++ src/client-module/dashboard/Sidebar.jsx
@@ -1,9 +1,9 @@
-import "../index.css";
+import "../../index.css";
 
 import PriceRange from "./PriceRangeContainer.jsx";
-import Categories from "./Categories";
+import Categories from "./Categories.jsx";
 
-const SideMenu = () => {
+const Sidebar = () => {
   return (
     <div className=" min-w-80 lg:w-1/3  bg-stone-200 hidden lg:block p-5 flex items-center justify-center">
       <PriceRange />
@@ -11,4 +11,4 @@ const SideMenu = () => {
     </div>
   );
 };
-export default SideMenu;
+export default Sidebar;
